requirements.txt
absl-py==0.13.0
apptools==5.1.0
astor==0.8.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
cached-property==1.5.2
cffi==1.14.5
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
dataclasses==0.8
envisage==6.0.1
gast==0.5.0
google-pasta==0.2.0
grpcio==1.38.1
h5py==3.1.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.6.1
importlib-resources==5.2.0
incremental==21.3.0
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
laspy==2.0.1
Markdown==3.3.4
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.3.1
pkg_resources==0.0.0
plyfile==0.7.4
protobuf==3.17.3
psutil==5.8.0
pycparser==2.20
pyface==7.3.0
Pygments==2.9.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.1
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tqdm==4.62.0
traits==6.2.0
traitsui==7.2.1
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.0
vtk==9.0.3
Werkzeug==2.0.1
winregistry==0.8.3
wrapt==1.12.1
wslink==0.2.0
zipp==3.5.0
zope.interface==5.4.0
